Feature/chat colors (#1806)
* Add tabbed sections to settings window.

* Add chat channel color customization to settings UI

* Incorporate OOC chat color and Basic into custom colors displayability.

* Fix emote style coloring

* Add emotes to text color options. Remove border around circles on UI.

+++ SWLOR.Game.Server/Service/GuiService/Component/GuiDrawListItemCircle.cs
@@ -47,7 +47,7 @@ namespace SWLOR.Game.Server.Service.GuiService.Component
         /// <param name="green">The amount of green to use. 0-255</param>
         /// <param name="blue">The amount of blue to use. 0-255</param>
         /// <param name="alpha">The amount of alpha to use. 0-255</param>
-        public GuiDrawListItemCircle<T> SetColor(int red, int green, int blue, int alpha = 255)
+        public GuiDrawListItemCircle<T> SetColor(byte red, byte green, byte blue, byte alpha = 255)
         {
             Color = new GuiColor(red, green, blue, alpha);
             return this;
